Stylistic improvements
Problem: there are several issues wuth css, and ContextActions interaction.

Solution: fix them.
