+++ frontend/src/pages/library/LibraryPage.tsx
@@ -1,29 +1,15 @@
 import React, { FunctionComponent, useState } from "react";
-import { useRecoilValue, useSetRecoilState } from "recoil";
 import cx from "classnames";
-import { Column } from "react-table";
-import { Link } from "react-router-dom";
-import { Table } from "../../components/Table/Table";
 import "./LibraryPage.scss";
 import { SuspenseSpinner } from "../../components/Spinner/SuspsenseSpinner";
 import PageLayout from "../../components/PageLayout/PageLayout";
 import { CreateMethodologyButton } from "../../components/buttons/CreateMethodologyButton";
 import { CreateProjectButton } from "../../components/buttons/CreateProjectButton";
-import { CompoundDto, MethodologyDto, ProjectDto, TargetDto } from "../../api/types";
-import { formatDateTimeDto } from "../../utils/utils";
 import { ErrorPlaceholder } from "../../components/Error/ErrorPlaceholder";
-import {
-  compoundsQuery,
-  methodologiesQuery,
-  projectsQuery,
-  targetsQuery,
-} from "../../store/selectors";
-import { ContextActions } from "../../components/ContextActions/ContextActions";
-import EditSvg from "../../assets/svg/edit.svg";
-import DeleteSvg from "../../assets/svg/delete.svg";
-import { Button } from "../../components/Button/Button";
-import { modalDialogAtom } from "../../store/atoms";
-import { EmptyPlaceholder } from "../../components/EmptyPlaceholder/EmptyPlaceholder";
+import { ProjectsSuspendable } from "./Projects";
+import { CompoundsSuspendable } from "./Compounds";
+import { MethodsSuspendable } from "./Methodologies";
+import { TargetsSuspendable } from "./Targets";
 
 export const LibraryPage: FunctionComponent = () => {
   return (
@@ -90,324 +76,5 @@ function EntitiesTab({ render, renderAddButton }: EntitiesTabProps) {
       </div>
       {render(entityTab)}
     </>
-  );
-}
-
-const extraFormatter = (compounds: string[]) => {
-  if (compounds.length <= 4) return compounds.join(", ");
-  return `${compounds.slice(0, 4).join(", ")} and ${compounds.length - 4} more`;
-};
-
-function ProjectsSuspendable() {
-  const setModalDialog = useSetRecoilState(modalDialogAtom);
-  const projects = useRecoilValue(projectsQuery);
-  const projectColumns = React.useMemo(
-    () => [
-      {
-        Header: "Project",
-        accessor: (p: ProjectDto) => p.item.name,
-      },
-      {
-        Header: "Compounds",
-        accessor: (p: ProjectDto) => (
-          <span className="project__compounds">{extraFormatter(p.item.compoundNames)}</span>
-        ),
-      },
-      {
-        Header: "Creation date",
-        accessor: (p: ProjectDto) => formatDateTimeDto(p.item.creationDate),
-      },
-      {
-        Header: "Last update",
-        accessor: (p: ProjectDto) => formatDateTimeDto(p.item.lastUpdate),
-      },
-      {
-        id: "actions",
-        accessor: (p: ProjectDto) => (
-          <ContextActions
-            actions={[
-              <div
-                key="edit"
-                className="contextActions__item"
-                onClick={() =>
-                  setModalDialog({
-                    kind: "create-edit-project",
-                    editing: p,
-                  })
-                }
-              >
-                <EditSvg />
-                Edit
-              </div>,
-            ]}
-          />
-        ),
-      },
-    ],
-    [setModalDialog]
-  );
-  return projects.length === 0 ? (
-    <EmptyPlaceholder
-      title="No projects created yet"
-      description="All created projects will be displayed here"
-      button={
-        <Button type="primary" onClick={() => setModalDialog({ kind: "create-edit-project" })}>
-          Create project
-        </Button>
-      }
-    />
-  ) : (
-    <Table data={projects} columns={projectColumns} />
-  );
-}
-
-function CompoundsSuspendable() {
-  const compounds = useRecoilValue(compoundsQuery);
-  const setModalDialog = useSetRecoilState(modalDialogAtom);
-  const compoundsColumns = React.useMemo(
-    () => [
-      {
-        Header: "Compounds",
-        accessor: (c: CompoundDto) => c.item.name,
-      },
-      {
-        Header: "MDe link",
-        // TODO form MDe link
-        accessor: (c: CompoundDto) => (
-          <a className="cellText" href={c.item.name}>
-            mde.io
-          </a>
-        ),
-      },
-      {
-        Header: "ChemSoft link",
-        id: "chemsoft",
-        accessor: (c: CompoundDto) => {
-          if (c.item.chemSoft)
-            return (
-              <td className="ednaTable__cell">
-                <a href={c.item.chemSoft}>{c.item.chemSoft}</a>
-              </td>
-            );
-          return (
-            <td className="ednaTable__cell cellBtn">
-              <Button
-                type="half-rounded"
-                size="small"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "add-edit-link",
-                    target: { kind: "compound", object: c },
-                  });
-                }}
-              >
-                Add link
-              </Button>
-            </td>
-          );
-        },
-      },
-      {
-        id: "actions",
-        accessor: (c: CompoundDto) => (
-          <ContextActions
-            actions={[
-              <div
-                key="edit"
-                className="contextActions__item"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "add-edit-link",
-                    target: { kind: "compound", object: c },
-                  });
-                }}
-              >
-                <EditSvg />
-                Edit
-              </div>,
-            ]}
-          />
-        ),
-      },
-    ],
-    [setModalDialog]
-  );
-  return compounds.length === 0 ? (
-    <EmptyPlaceholder
-      title="No compounds added yet"
-      description="To add compounds add new experiments"
-      button={
-        <Link to="/upload" className="libraryPage__linkButton">
-          <Button type="primary">Add experiments</Button>
-        </Link>
-      }
-    />
-  ) : (
-    <Table
-      data={compounds}
-      columns={compoundsColumns}
-      columnExtras={{
-        chemsoft: { manualCellRendering: true },
-      }}
-    />
-  );
-}
-
-function TargetsSuspendable() {
-  const targets = useRecoilValue(targetsQuery);
-  const targetColumns: Column<TargetDto>[] = React.useMemo(
-    () => [
-      {
-        Header: "Target",
-        accessor: (t: TargetDto) => t.item.name,
-      },
-      {
-        Header: "Projects",
-        accessor: (t: TargetDto) => extraFormatter(t.item.projects),
-      },
-      {
-        Header: "Addition date",
-        accessor: (t: TargetDto) => formatDateTimeDto(t.item.additionDate),
-      },
-    ],
-    []
-  );
-  return targets.length === 0 ? (
-    <EmptyPlaceholder
-      title="No targets added yet"
-      description="To add targets add new experiments"
-      button={
-        <Link to="/upload" className="libraryPage__linkButton">
-          <Button type="primary">Add experiments</Button>
-        </Link>
-      }
-    />
-  ) : (
-    <Table data={targets} columns={targetColumns} />
-  );
-}
-
-function MethodsSuspendable() {
-  const setModalDialog = useSetRecoilState(modalDialogAtom);
-  const methodologies = useRecoilValue(methodologiesQuery);
-  const methodologyColumns: Column<MethodologyDto>[] = React.useMemo(
-    () => [
-      {
-        Header: "Methodology",
-        accessor: (t: MethodologyDto) => t.item.name,
-      },
-      {
-        Header: "Project",
-        accessor: () => "project1, project2",
-      },
-      {
-        Header: "Confluence link",
-        id: "link",
-        accessor: (m: MethodologyDto) => {
-          if (m.item.confluence)
-            return (
-              <td className="ednaTable__cell">
-                <a href={m.item.confluence}>{m.item.confluence}</a>
-              </td>
-            );
-          return (
-            <td className="ednaTable__cell libraryTable__cellBtn">
-              <Button
-                type="half-rounded"
-                size="small"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "add-edit-link",
-                    target: { kind: "methodology", object: m },
-                  });
-                }}
-              >
-                Add link
-              </Button>
-            </td>
-          );
-        },
-      },
-      {
-        Header: "Description",
-        id: "description",
-        accessor: (m: MethodologyDto) => {
-          return (
-            <td className="ednaTable__cell cellBtn">
-              <Button
-                type="half-rounded"
-                size="small"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "methodology-description",
-                    methodology: m,
-                  });
-                }}
-              >
-                Show description
-              </Button>
-            </td>
-          );
-        },
-      },
-
-      {
-        id: "actions",
-        accessor: (m: MethodologyDto) => (
-          <ContextActions
-            actions={[
-              <div
-                key="edit"
-                className="contextActions__item"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "create-edit-methodology",
-                    editing: m,
-                  });
-                }}
-              >
-                <EditSvg />
-                Edit
-              </div>,
-              <div
-                key="delete"
-                className="contextActions__item"
-                onClick={() => {
-                  setModalDialog({
-                    kind: "delete-methodology",
-                    methodology: m,
-                  });
-                }}
-              >
-                <DeleteSvg />
-                Delete
-              </div>,
-            ]}
-          />
-        ),
-      },
-    ],
-    [setModalDialog]
-  );
-
-  return methodologies.length === 0 ? (
-    <EmptyPlaceholder
-      title="No methodologies created yet"
-      description="All created methodologies will be displayed here"
-      button={
-        <Button type="primary" onClick={() => setModalDialog({ kind: "create-edit-methodology" })}>
-          Create methodology
-        </Button>
-      }
-    />
-  ) : (
-    <Table
-      data={methodologies}
-      columns={methodologyColumns}
-      columnExtras={{
-        link: { manualCellRendering: true },
-        description: { manualCellRendering: true },
-      }}
-    />
   );
 }
